Refactor: 어드민 게시물 작성 시 데이터 형식, 이미지 형식 관련 유틸 함수 분리
