+++ src/app/admin/notice/[noticeId]/edit/page.tsx
@@ -9,59 +9,16 @@ import * as styles from '../../create/page.css';
 
 import updateNotice from '@/app/_api/notice/updateNotice';
 import uploadNoticeImages from '@/app/_api/notice/uploadNoticeImages';
+import getNoticeDetail from '@/app/_api/notice/getNoticeDetail';
 
 import { NoticeCreateType, NoticeDetailType } from '@/lib/types/noticeType';
 import { noticeDescriptionRules, noticeTitleRules } from '@/lib/constants/formInputValidationRules';
+import { QUERY_KEYS } from '@/lib/constants/queryKeys';
+import { NOTICE_CATEGORY_NAME } from '@/lib/constants/notice';
+import { formatImageData, formatNoticeData } from '@/lib/utils/formatDataForNotice';
 
 import CategoryDropdown from '../../create/_components/CategoryDropdown';
 import ContentsBody from '../../create/_components/ContentsBody';
-import getNoticeDetail from '@/app/_api/notice/getNoticeDetail';
-import { QUERY_KEYS } from '@/lib/constants/queryKeys';
-import { NOTICE_CATEGORY_NAME } from '@/lib/constants/notice';
-
-/** 공지 작성 데이터 포맷 유틸 함수 */
-const formatNoticeData = (originData: NoticeCreateType) => {
-  // order 정리 및 이미지 url 초기화
-  const updatedContents = originData.contents.map((item, index) => {
-    const newContents = { ...item, order: index + 1 };
-    if (newContents.type === 'image') {
-      if (typeof newContents.imageUrl !== 'string') {
-        newContents.imageUrl = '';
-      }
-    }
-    return newContents;
-  });
-
-  // 새로운 데이터 객체 반환
-  const noticeData: NoticeCreateType = {
-    ...originData,
-    contents: updatedContents,
-  };
-  return noticeData;
-};
-
-/** 이미지 업로드 데이터 포맷 유틸 함수 */
-const formatImageData = (originData: NoticeCreateType) => {
-  const imageExtensionData = originData.contents
-    .map((item, index) => {
-      return { ...item, order: index + 1 };
-    })
-    .filter((item) => item.type === 'image' && typeof item.imageUrl !== 'string')
-    .map(({ order, imageUrl }) => {
-      return {
-        order: order,
-        extension: (imageUrl as File).type.split('/')[1],
-      };
-    });
-
-  const imageFileData = originData.contents
-    .filter((item) => item.type === 'image' && typeof item.imageUrl !== 'string')
-    .map((item) => {
-      return item.imageUrl as File;
-    });
-
-  return { imageExtensionData, imageFileData };
-};
 
 type NoticeCategoryNameType = (typeof NOTICE_CATEGORY_NAME)[keyof typeof NOTICE_CATEGORY_NAME];
