+++ src/app/admin/notice/create/page.tsx
@@ -12,53 +12,10 @@ import uploadNoticeImages from '@/app/_api/notice/uploadNoticeImages';
 
 import { NoticeCreateType } from '@/lib/types/noticeType';
 import { noticeDescriptionRules, noticeTitleRules } from '@/lib/constants/formInputValidationRules';
+import { formatImageData, formatNoticeData } from '@/lib/utils/formatDataForNotice';
 
 import CategoryDropdown from './_components/CategoryDropdown';
 import ContentsBody from './_components/ContentsBody';
-
-/** 공지 작성 데이터 포맷 유틸 함수 */
-const formatNoticeData = (originData: NoticeCreateType) => {
-  // order 정리 및 이미지 url 초기화
-  const updatedContents = originData.contents.map((item, index) => {
-    const newContents = { ...item, order: index + 1 };
-    if (newContents.type === 'image') {
-      if (typeof newContents.imageUrl !== 'string') {
-        newContents.imageUrl = '';
-      }
-    }
-    return newContents;
-  });
-
-  // 새로운 데이터 객체 반환
-  const noticeData: NoticeCreateType = {
-    ...originData,
-    contents: updatedContents,
-  };
-  return noticeData;
-};
-
-/** 이미지 업로드 데이터 포맷 유틸 함수 */
-const formatImageData = (originData: NoticeCreateType) => {
-  const imageExtensionData = originData.contents
-    .map((item, index) => {
-      return { ...item, order: index + 1 };
-    })
-    .filter((item) => item.type === 'image' && typeof item.imageUrl !== 'string')
-    .map(({ order, imageUrl }) => {
-      return {
-        order: order,
-        extension: (imageUrl as File).type.split('/')[1],
-      };
-    });
-
-  const imageFileData = originData.contents
-    .filter((item) => item.type === 'image' && typeof item.imageUrl !== 'string')
-    .map((item) => {
-      return item.imageUrl as File;
-    });
-
-  return { imageExtensionData, imageFileData };
-};
 
 export default function CreateNotice() {
   const router = useRouter();
